# swagger.json
{
    "openapi": "3.0.0",
    "paths": {
        "/auth/login": {
            "post": {
                "operationId": "AuthController_login",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/roles": {
            "post": {
                "operationId": "AuthController_createRole",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateRoleDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "defaultBearerAuth": []
                    }
                ]
            },
            "get": {
                "operationId": "AuthController_getRoles",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "defaultBearerAuth": []
                    }
                ]
            }
        },
        "/auth/users/role": {
            "post": {
                "operationId": "AuthController_addRole",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AddRoleDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "defaultBearerAuth": []
                    }
                ]
            }
        },
...
        "/film": {
            "post": {
                "operationId": "FilmController_createFilm",
                "summary": "Создание фильма",
                "parameters": [],
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Promise"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "film"
                ]
            },
            "get": {
                "operationId": "FilmController_getAll",
                "summary": "Получение информации о всей кинопродукции",
                "parameters": [
                    {
                        "name": "genres",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "202": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Film"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "film"
                ]
            }
        },
        "/film/movies": {
            "get": {
                "operationId": "FilmController_getAllMovies",
                "summary": "Получение информации о фильмах",
                "parameters": [
                    {
                        "name": "genres",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "202": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Film"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "film"
                ]
            }
        },
        "/film/serials": {
            "get": {
                "operationId": "FilmController_getAllSerials",
                "summary": "Получение информации о сериалах",
                "parameters": [
                    {
                        "name": "genres",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "202": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Film"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "film"
                ]
            }
        },
        "/film/cartoons": {
            "get": {
                "operationId": "FilmController_getAllCartoons",
                "summary": "Получение информации о мультфльмах",
                "parameters": [
                    {
                        "name": "genres",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "202": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Film"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "film"
                ]
            }
        },
        "/film/{id}": {
            "get": {
                "operationId": "FilmController_getFilm",
                "summary": "Получение информации о конкретном фильме",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "202": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Film"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "film"
                ]
            },
            "patch": {
                "operationId": "FilmController_updateFilm",
                "summary": "Обновление информации о фильме",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateFilmDTO"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "202": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Promise"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "film"
                ]
            },
            "delete": {
                "operationId": "FilmController_deleteFilm",
                "summary": "Удаление фильма",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "202": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteResult"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "film"
                ]
            }
        },
...
    },
    "info": {
        "title": "Учебный проект",
        "description": "Клон kinopoisk-a",
        "version": "0.0.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "defaultBearerAuth": {
                "scheme": "bearer",
                "bearerFormat": "jwt",
                "type": "http"
            }
        },
        "schemas": {
            "LoginDto": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "user@mail.ru",
                        "description": "Почтовый адрес"
                    },
                    "password": {
                        "type": "string",
                        "minLength": 4,
                        "maxLength": 16,
                        "example": "12345678",
                        "description": "Пароль пользователя"
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
            "CreateRoleDto": {
                "type": "object",
                "properties": {
                    "value": {
                        "type": "string",
                        "example": "Администратор",
                        "description": "Роль пользователя"
                    },
                    "description": {
                        "type": "string"
                    }
                },
                "required": [
                    "value",
                    "description"
                ]
            },
            "AddRoleDto": {
                "type": "object",
                "properties": {
                    "value": {
                        "type": "string",
                        "example": "Администратор",
                        "description": "Роль пользователя"
                    },
                    "userId": {
                        "type": "number",
                        "minimum": 1
                    }
                },
                "required": [
                    "value",
                    "userId"
                ]
            },
...
            "Promise": {
                "type": "object",
                "properties": {}
            },
...
            "ActorFilm": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "minimum": 1,
                        "description": "Первичный ключ",
                        "example": 1
                    },
                    "filmId": {
                        "type": "number",
                        "minimum": 1,
                        "description": "Foreign key",
                        "example": 1
                    },
                    "actorId": {
                        "type": "number",
                        "minimum": 1,
                        "description": "Foreign key",
                        "example": 1
                    },
                    "roleId": {
                        "type": "number",
                        "minimum": 1,
                        "description": "Foreign key",
                        "example": 1
                    },
                    "roleNotes": {
                        "type": "string",
                        "description": "Примечание по роли",
                        "examples": [
                            "... дворецкий",
                            "... по костюмам"
                        ]
                    },
                    "film": {
                        "$ref": "#/components/schemas/Film"
                    },
                    "actor": {
                        "$ref": "#/components/schemas/Actor"
                    },
                    "role": {
                        "$ref": "#/components/schemas/ActorRole"
                    }
                },
                "required": [
                    "id",
                    "filmId",
                    "actorId",
                    "roleId",
                    "roleNotes",
                    "film",
                    "actor",
                    "role"
                ]
            },
            "FilmGenre": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "minimum": 1,
                        "description": "Первичный ключ",
                        "example": 1
                    },
                    "genreId": {
                        "type": "number",
                        "minimum": 1,
                        "description": "Foreign key",
                        "example": 1
                    },
                    "filmId": {
                        "type": "number",
                        "minimum": 1,
                        "description": "Foreign key",
                        "example": 1
                    },
                    "film": {
                        "$ref": "#/components/schemas/Film"
                    }
                },
                "required": [
                    "id",
                    "genreId",
                    "filmId",
                    "film"
                ]
            },
...
            "Film": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "minimum": 1,
                        "description": "Первичный ключ",
                        "example": 1
                    },
                    "url": {
                        "type": "string",
                        "description": "Ссылка фильм на кинопоиске",
                        "example": "https://www.kinopoisk.ru/film/301/"
                    },
                    "preview": {
                        "type": "string",
                        "description": "Ссылка на preview"
                    },
                    "year": {
                        "type": "number",
                        "minimum": 1,
                        "description": "Год выпуска",
                        "example": 1999
                    },
                    "title": {
                        "type": "string",
                        "description": "Название фильма",
                        "example": "Матрица"
                    },
                    "originalTitle": {
                        "type": "string",
                        "description": "Оригинальное название",
                        "example": "Matrix"
                    },
                    "slogan": {
                        "type": "string",
                        "description": "Слоган",
                        "example": "«Добро пожаловать в реальный мир»"
                    },
                    "countryId": {
                        "type": "number",
                        "minimum": 1,
                        "description": "Foreign Key",
                        "example": 1
                    },
                    "ageRestrictionId": {
                        "type": "number",
                        "minimum": 1,
                        "description": "Foreign Key",
                        "example": 1
                    },
                    "country": {
                        "$ref": "#/components/schemas/Country"
                    },
                    "personsInFilm": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ActorFilm"
                        }
                    },
                    "filmGenres": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FilmGenre"
                        }
                    },
                    "duration": {
                        "type": "string"
                    },
                    "ageRestriction": {
                        "$ref": "#/components/schemas/AgeRestriction"
                    },
                    "reviews": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Review"
                        }
                    }
                },
                "required": [
                    "id",
                    "url",
                    "year",
                    "title",
                    "originalTitle",
                    "slogan",
                    "countryId",
                    "ageRestrictionId",
                    "country",
                    "personsInFilm",
                    "filmGenres",
                    "duration",
                    "ageRestriction",
                    "reviews"
                ]
            },
...
            "Country": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number",
                        "minimum": 1,
                        "description": "Первичный ключ",
                        "example": 1
                    },
                    "countryName": {
                        "type": "string",
                        "description": "Название страны",
                        "example": "USA"
                    },
                    "url": {
                        "type": "string",
                        "description": "Ссылка на страну на кинопоиске",
                        "example": "https://www.kinopoisk.ru/lists/movies/country--1/?b=films&b=top"
                    },
                    "filmsCreated": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Film"
                        }
                    },
                    "filmViews": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FilmViewsCountry"
                        }
                    }
                },
                "required": [
                    "id",
                    "countryName",
                    "url",
                    "filmsCreated",
                    "filmViews"
                ]
            },
            "CreateFilmDTO": {
                "type": "object",
                "properties": {
                    "year": {
                        "type": "number",
                        "minimum": 1,
                        "description": "Год выпуска",
                        "example": 1999
                    },
                    "title": {
                        "type": "string",
                        "description": "Название фильма",
                        "example": "Матрица"
                    },
                    "originalTitle": {
                        "type": "string",
                        "description": "Оригинальное название",
                        "example": "Matrix"
                    },
                    "slogan": {
                        "type": "string",
                        "description": "Слоган",
                        "example": "«Добро пожаловать в реальный мир»"
                    },
                    "countryName": {
                        "type": "string",
                        "description": "Назвaние страны",
                        "example": "USA"
                    },
                    "duration": {
                        "type": "string"
                    }
                },
                "required": [
                    "year",
                    "title",
                    "originalTitle",
                    "slogan",
                    "countryName",
                    "duration"
                ]
            },
            "DeleteResult": {
                "type": "object",
                "properties": {}
            },
...
        }
    }
}